+++ src/Websites/Websites.Test/ScenarioTests/WebAppTests.ps1
@@ -1285,6 +1285,46 @@ function Test-PublishAzureWebAppOnedeploy
 	}
 }
 
+# New tests for PublishAzureWebApp to test deploying against ZipDeploy
+function Test-PublishAzureWebAppZipDeploy
+{
+	# Setup
+	$rgname = Get-ResourceGroupName
+	$appName = Get-WebsiteName
+	$location = Get-WebLocation
+	$planName = Get-WebHostPlanName
+	$tier = "Shared"
+
+	try
+	{
+		#Setup
+		New-AzureRmResourceGroup -Name $rgname -Location $location
+		$serverFarm = New-AzureRmAppServicePlan -ResourceGroupName $rgname -Name  $planName -Location  $location -Tier $tier
+		
+		# Create new web app
+		$webapp = New-AzureRmWebApp -ResourceGroupName $rgname -Name $appName -Location $location -AppServicePlan $planName 
+		
+		#Configuring jdk and web container
+        # Set Java runtime to 1.8 | Tomcat. In order to deploy war, site should be configured to run with stack = TOMCAT 
+		# or JBOSSEAP (only availble on Linux). In this test case, it creates Windows app. 
+		$javaVersion="1.8"
+        $javaContainer="TOMCAT"
+        $javaContainerVersion="8.5"
+        $PropertiesObject = @{javaVersion = $javaVersion;javaContainer = $javaContainer;javaContainerVersion = $javaContainerVersion}
+        New-AzResource -PropertyObject $PropertiesObject -ResourceGroupName $rgname -ResourceType Microsoft.Web/sites/config -ResourceName "$appName/web" -ApiVersion 2018-02-01 -Force
+
+		$warPath = Join-Path $ResourcesPath "HelloJava.war"
+		$publishedApp = Publish-AzWebApp -ResourceGroupName:$rgname -Name:$appName -ArchivePath:$warPath -UseZipDeploy -Force
+
+		Assert-NotNull $publishedApp
+	}
+	finally
+	{
+		# Cleanup
+		Remove-AzureRmResourceGroup -Name $rgname -Force
+	}
+}
+
 <#
 .SYNOPSIS
 Tests creating a web app with a simple parameterset.